requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-ninja==1.2.2
django-role-permissions==3.2.0
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
